Change "Home" link to "Courses" when user has no active enrollments & groups
When the current user has no active enrollments or groups the Courses dropdown changes to a Home link. In doing so there is no way to get to the Courses page to see prior enrollments without knowing the direct URL. This mod changes the Home link to a link to /courses.

Test plan:
1. Log in as a user with no active enrollments (see me for a test user ID)
2. Verify that Courses is in the top nav bar and is a link to /courses