Add "sections" property to the Amaint course API
This is part of our implementation of multiple enrollments, in which
students are enrolled in *both* lecture and tutorial/lab sections.

The "sections" property returns all sections in a given course, and
replaces the older "sectionTutorials" property.

NOTE: In the event that the REST API call returns a 404 or 500 (or more
accurately, if the JSON can't be parsed as an Array), an exception will
be raised and the output will be a standard Canvas error JSON. Before,
it would also have resulted in an error JSON, but the failure will occur
at a different, less obvious place: inside #is_enrollment_section?
because #any? is called on an unexpected number (json_data).

Test plan:
* GET /sfu/api/v1/amaint/course/<course_code>/sections
* Verify that the JSON result has "sections" that lists all sections