remove `include_root_in_json_code`
We previously did some shenanigans (a3cf39) to remove the root element from the SFU terms API. This seems to have broken in a recent deploy (or possibly as far back as Rails 3, who knows). AEM is the main consumer of this API, and Ron already changed his code to expect the root element, so we're not "fixing" it. This simply removes the `include_root_in_json` calls that aren't doing anything.