Move setting persistent js_env values into ApplicationController#js_env
Instructure changed how js_env operates in ddf6184 and it broke our monkeypatch. This moves the setting of our persistent js_env values into the main ApplicationControlller#js_env method, wrapped in a SFU MOD. This is the less-dirty approach and has the advantage of likely throwing a conflict if Instructure changes something in there again.