Really remove dashboard sidebar SFU MOD in users_controller
We previously "removed" this mod in b9d496d, but didn't fully remove it: the instructure code was restored, but our actual mod (below the cache code) was kept. This caused issues fetching the sidebar after the deploy on 2016-02-02.

This commit restores the file to the instructure version as at the 2016-01-09 release. Already hotfixed in place on production.