Add sfu_help plugin
Overrides the default Canvas::Help.default_links method to return an
empty array. When Account#help_links is called, it'll merge the custom
links with the empty array.

Implemented as a simple Railtie (thanks @ccutrer). Putting it inside the
main repo instead of on its own because it's so stupidly simple.