Remove use of sis_source_id and unused SFU MOD
Instructure removed the sis_source_id column from the Enrollment model
in e8ee04b. This caused a 500 ISE in the "List users in course" API
endpoint, which also broke the Gradebook.

This is safe to remove because CANVAS-224 was phased out as of the
2015-12-19 release; it's an unused SFU MOD.

(Also, add SFU MOD tag to the next line.)