Remove unused @numberOfFrozenCols from Gradebook
`@numberOfFrozenCols` was removed in an upstream change:

  > instructure/canvas-lms@62db47a269d68126c4e7183effbddc0d55285b20

Our SFU MOD to modify `@numberOfFrozenCols` is no longer needed because
the frozen column count is now determined by the length of
`@parentColumns`, which is where we have always added the SIS ID column.

This commit should not result in any functional changes.