TEMP: Disable features using `require` in sfu.js
RequireJS has been replaced with a shim that only allows certain whitelisted bundles to load; everything else is supposed to use the new webpack build stuff. This is breaking two items in sfu.js: the addition of the copyright modal dialog when a course is published, and the insertion of the privacy notice on the Google Docs collaboration page. I have temporarily disabled those two features until a solution can be found.