Statically set api_max_per_page in new gradebook to 50
We overwrite the API max per page limit elsewhere to make the enrollments scripts more efficient. The new gradebook attempts to fetch the roster in batches based on per-page limits. Because ours is so high, it essentially attempts to fetch the entire course at once by building an absurdly-long URL, one which on large courses will be bigger than Apache's limit. This mod explicitly sets it to the original default, 50 for the new gradebook call only.