Use Account `sis_source_id` to build course CSV
The course form was using the `id` rather than `sis_source_id` of the
default account. This worked in the past because SIS::CourseImporter
will use the default (root) account anyway. But now that it checks the
account's presence (f66372aab250843185754e3eede6438a41e23361), an
invalid `account_id` in the SIS CSV will cause course creation to fail.

Note that this fix will still work even if the default account does not
have an `sis_source_id` (nil), because in that case SIS::CourseImporter
will skip that check and continue using the default account as usual.

Test plan:

1. Create a credit course using the SFU Course Form
2. The course should be created successfully in the SFU account